requirements.txt
django-revproxy
django-uuidfield
requests